Refactor KeyContext to implement the whole Destroyable interface
Why:
It was inconsistent to half implement it and also have a similar method
isAlive. Move the encryptor destruction try/catch into the KeyContext
as it's a concern of that class.
